requirements.txt
flask
flask_sqlalchemy
flask_wtf
wtforms[email]
flask-sijax
flask_socketio
flask_migrate
flask_bcrypt
flask_jwt_extended
config
pytest-order
pytz